requirements.txt
# Core dependencies
google-generativeai>=0.8.0
langchain>=0.1.0
langchain-core>=0.1.0
langchain-tavily>=0.2.0
langgraph>=0.0.50
langgraph-checkpoint>=0.0.10
langgraph-prebuilt>=1.0.0
langgraph-sdk>=0.3.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.0.0
typing-extensions>=4.0.0

# Development dependencies (optional)
pytest>=7.0.0
pytest-asyncio>=0.21.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.0.0
